requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
flake8==6.1.0
autoflake==2.2.1
isort==5.12.0
pytest==7.4.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.1

